requirements.txt
python-socketio
aiohttp
pygame
aiohttp_cors